+++ src/app/order/store/order.actions.ts
@@ -1,0 +1,4 @@
+import {createAction, props} from "@ngrx/store";
+import {CartItemModel} from "../../cart/models/cart-item.model";
+
+export const createOrder = createAction('[Order] Create order', props<{products: CartItemModel[]}>());