+++ src/app/store/order.selectors.ts
@@ -1,0 +1,3 @@
+import {AppState} from "./app.reducer";
+
+export const selectOrderState = (state: AppState) => state.order;