+++ changelog.md
@@ -4,6 +4,10 @@
 ### Added
 * Added NgRx for Product and Cart module
 * Added feature and state selectors for product and cart states
+* Added Router/Store
+* Added product caching with the store
+* Added router store
+* Added navigation to the order page via router action
 
 ## Task 6. HttpClient
 ### Added